bugfix: double render still frames for SDL 2.0.1 on Mac OS X 10.9
If a still frame is only drawn once, the swapped back buffer is not redrawn,
resulting in violent flickering of the screen.  The only documentation I can
find on SDL double rendering is on SDL_GLattr, which maybe means that Mac
defaults to using an OpenGL renderer with back buffering since we don't specify
anything else.

This patch causes still frames to be drawn a second time before spinning down
the rendering pipeline.