Manually unwind gosh dang everything.
Since Python does not provide a reliable destructor for classes, we have to keep track of when we're done using SDL stuff and free it ourselves, when possible. Otherwise we continue to make a best effort with __del__. This solves #70 and #78, but introduces a peculiar issue on line 129 of areamanager.py, which has been commented out while it is being investigated. Everything should work regardless though there may be a resulting memory leak.