Clear every frames to black before drawing
AreaManager._tick() is the first function called every frame in
Driftwood's drawing process. If there's been a refocusing then
AreaManager asks FrameManager to prepare a new frame. This new frame is
black. Otherwise, the old frame is used and is not cleared, leaving
garbage on-screen if the current draw does not overwrite every pixel.

We need to clear the frame really early in the drawing process,
basically before AreaManager.__build_frame() is called since that's
where our drawing starts.

This commit adds a FrameManager tick that is called before AreaManager's
tick. All it does is clear the screen.

Fixes #75