Work on docs, Makefile updates.
More work on documentation. Moved documentation source to docsrc/, and it now builds into docs/. Since compiling Python bytecode makes it only safe to run on the platform on which it was compiled, we now have the regular make command in addition to a "make release" which creates the driftwood pyz executable out of the uncompiled source files and places it in the release/ folder. The Makefile now handles building the documentation and checking if it needs to be rebuilt.