AreaManager: Clip the top of entities to the area
When the screen is rumbling, all area graphics are drawn at an offset
of their normal locations. However, this can allow entities to be drawn
outside of the area if part of their graphic extends out of the area.
This shouldn't happen.

This commit crops the top of entities to fix the area.

Fixes #75.