Decide to collide the center of an entity with a tile.
A tile collision happens once the center of the entity is touching the edge of the tile. Alternatively, you could look at it as the edge of the entity colliding with the center of the tile. Also remove the mysterious and totally unused px and py arguments from _tile_at() and _tile_cross().