AreaManager: Push srcrect & dstrect calcs up
Bring the Tile.srcrect() and Tile.dstrect() calculations up into
the body of AreaManager.__build_frame().

This is preparation for the next step, which is predicting what a tile's
dstrect will be and skipping tiles that will be offscreen.

We also save ~11,880 allocations per second by reusing the srcrect and
dstrect lists between tiles. (Actual amount varies by viewport size and
client FPS.)