Consolidate schemas and remove base module concept.
PyInstaller, which we use to compile Windows binaries, doesn't like loading schemas from inside a pyz, and it doesn't like sharing the stdlib from inside the executable. We can improve compatibility, make releases easier, and save some code by consolidating the JSON schemas into variables in a python file, and moving the stdlib into the data directory. Now instead of a base module, the path manager puts a "common package" at the top of the path list, which contains the stdlib and and other common data we might want every world to have access to. Currently the common package is stored at "data/__common__/".